MessageGroupSubscription: Use groupId when queuing messages
In Idf55b3d6e0643d02f8bf8f281b384e1b802a786b, we determined the group
id of messages based on message index but messages for which we want
to send notification may not yet be in the message index.

Pass the group id explicitly when queuing the message for notification.

Bug: T375376
Change-Id: If6a7e4d8810cb1c803a2da4fd2e79e66c4e7c065