Change <languages/> tag to only use cached stats
Revision parsers after marking a page for translation are causing
problems as described in T369602. Hence updating TranslatablePage
::getTranslationPercentages to only use cached stats to avoid
causing slow stats rebuilds at unfortunate times. This method is
currently only used by the languages tag and the page header, so
it should be okay. In addition every time page is marked for
translation, jobs are spawned to update or purge all related pages,
making the system self-correcting in case a stale or missing value
is used.

Bug: T369602
Change-Id: Ia5f52821ee3c87c3662ccab2072d9074e7301a0a