Lessify LESS files, part 1
Turn the LESS files that were converted from CSS to LESS
into proper LESS. In this step I am only taking advantage
of LESS's nesting features – the CSS it outputs should mostly
be the same as it was (except some times in
a different order).

Split this into several patches because some of the changes
will be very large.

Bug: T371071
Change-Id: I5e67eaf48db05f103a136980de5f5329428b62c3