Lessify LESS files, part 2
Turn the LESS files that were converted from CSS to LESS
into proper LESS. In this step I am only taking advantage
of LESS's nesting features – the CSS it outputs should be
almost exactly the same as it was (except some times in
a different order).

Split into several patches because some of the changes
will be very large.

Bug: T371071
Change-Id: Ie8f2ca607565cc2a4c1f3675069befacc85ebf85