Lessify LESS files, part 3
Turn the LESS files that were converted from CSS to LESS
into proper LESS. In this step I am only taking advantage
of LESS's nesting features – the CSS it outputs should be
almost exactly the same as it was (except some times in
a different order).

Split into several patches because some of the changes
will be very large.

Bug: T371071
Change-Id: I814c2659824325c9a19200a3a53f052921265bce