Special:Translate: Use Codex icons
In preparation for dark mode support, exchange "custom" icons
with standard icons from Codex as much as possible. Most of the
"custom" icons were idential or nearly identical with the Codex
icons they're being replaced with, so for the most part, there
is little change in practice.

There are, however, some differences:
* The Codex icon for a checkmark is different (different angle)
* The Codex icon for edit (pen) is different (different angle,
  Codex icon is "bolder")
* The Codex icon for paste is very different (old icon was a
  clipboard, Codex icon is the copy icon with a plus on it)
* The clock/diff icon is solid in Codex, but an outline in the
  "custom" style.

Some icons don't have any equivalents in Codex; for those, I
have used the approach of using `.cdx-mixin-css-icon()` with
`'none'` as the first parameter, then the SVG file as
`mask-image`, as recommended in Codex's `css-icon.less`.

Bug: T381834
Change-Id: Iae9ed91fe2cb3a41460b1c27273d1c4963faf443