Special:AggregateGroups: Use Codex icons
In preparation for dark mode support, exchange "custom" icons
with standard icons from Codex as much as possible. In this
change, I've exchanged colorful SVG icons with plainer, but
more standardized Codex icons.

The "Depends-On" is because some of the icons deleted in this
patch are in use in Special:Translate, so if this is merged
before the other patch, those icons would stop appearing in
Special:Translate. After the two patches are both merged, none
of these icons will be in use any more.

Bug: T381834
Depends-On: Iae9ed91fe2cb3a41460b1c27273d1c4963faf443
Change-Id: Id4b5dda37a1b34d30c9196a679c71c1ef8683e21