Special:ManageTranslatorSandbox: Use Codex icons
In preparation for dark mode support, exchange "custom" icons
with standard icons from Codex as much as possible. In this patch,
only a search icon is replaced.

The "Depends-On" is because the search icon is also used in
Special:Translate, so if this patch is merged before the other one,
that icon would stop appearing in Special:Translate. After the two
patches are both merged, the file won't be in use any more.

Bug: T381834
Depends-On: Iae9ed91fe2cb3a41460b1c27273d1c4963faf443
Change-Id: Ifb8edd668a131dbcfb125950ac5812c484780b95