feat: Add dialog to add/edit aggregate groups
Added a dialog box to create or modify aggregate groups. The plan is
to refresh the page when a aggregate group is added or edited.
This allows the aggregate group generation code to be present only
on the server side.

This does make adding new aggregate groups on wikis that have a large
number of aggregate groups slower but addition/updating of aggregate
groups is not a frequent occurence.

Bug: T382949
Change-Id: Ia9f636f537b104d166498d476cc19402234d37d0