Remove default 'en' language code for aggregate groups
Aggregate message groups may or may not have a source language set.
The code in MessageGroupBase returns 'en' if no source language is
set. This is fine for FileBasedMessageGroup but for
AggregateMessageGroups return 'und' if no source language is
specified.

For AggregateGroups that are defined in YAML configuration files
default to using wiki content language if no source language is
specified.

Motivation: In I3b742839ac0de7540875db1c45c5bed786497423, we fetch
aggregate message group before editing them. If source language
falls back to 'en' when a source language code is not set then the
incorrect source language is displayed on the edit form.

Change the code in the dialog to account for the und value.

Bug: T382949
Change-Id: I1b76a6d698ff60215a4871a3a9b627dceae8ced6