Special:ManageTranslatorSandbox: Use skin variables for dark mode support
Use skin variables where it makes sense, in order to support dark
mode for this special page.

There are a few visual differences – mostly just different shades of
grey; the most noticeable difference is the hover and selected state
for the request list, where the (semantically) best matches from Codex
are quite a bit darker than what was here before. I also removed the
different-colored background for the number of translations in that
list – since it already has a comparatively huge font, I don't
think we need yet another thing there to visually separate it from
the username portion of the request.

Bug: T381834
Change-Id: I3140d2132a353fa62465524fdf7342e48710a95f