feat: Display initial suggestions in lookup on aggregate group refresh
Display a few initial suggestion on the message group lookup on
aggregate group refresh page.

Codex Lookup has some weird behavior where it does not detect changes
to menu items unless it is initialized with an empty object.

Also add a request tracker to avoid making multiple API calls to the
entitysearch API. Without this an API call would be made for every
aggregate group displayed on the special page.

Bug: T382949
Change-Id: I9a4cdce482c4dd3d5e0367f38b8e6218dce2fc7c