Remove ineffective TOC-setting code
In the process of attempting to update this code to match the revised
function signature for OutputPage::addParserOutput(), Tacspipacsi
noticed that the `$pout->setTOCData(...)` line that most other special
pages have here is missing.  That is, this is adding a TOC placeholder
but there's no actual TOC data present here, meaning this codes isn't
doing anything at all.

Bug: T350626
Change-Id: I6b337d673d3c98c532e8a5da650836aa3d4d6b51