ActiveLanguagesSpecialPage: Use even more PHP 8 language constructs
Not only promoted properties can be readonly, but also properties
normally set in the constructor. (And even properties like
`$progressStatsTable` that aren’t set in the constructor,
as long as they aren’t assigned more than once. But Phan doesn’t support
this, and there is no consensus on using it, so don’t do it.)

Reorder properties so that the readonly (set-in-constructor) properties
are grouped together.

`$period` used to be initialized to a constant value and never changed,
so turn it into a constant, as suggested by the readonly properties RFC
(https://wiki.php.net/rfc/readonly_properties_v2#restrictions).

Also update a core class namespaced in 1.44.

Follows-Up: 72ae216838ad22eb2e9749bd89f668b57cb723b2
Change-Id: Ided8e951b8ef94c3145f765f5277f2c2d59e233a