Change wording of tpt-select-prioritylangs message
The list it references is no longer a list of language codes,
nor is it comma-separated. Instead add a help text that explains
the format is JavaScript is disabled.

Convert the help to inline format for better accessiblity and
display it only when JavaScript is disabled.
(Niklas): In my testing clicking the help icon does nothing,
tooltip is shown on hover.

Also stop parsing the other help text, no reason given when it
was introduced.

Follows-up: 7f9bd8c26dbbefcb675d5f529b624b9fd50cb2ee
Change-Id: Ia849904a9afdcd1ae156364aaa4a46922f9633a1