feat: Add links to Special:Translations from Special:PageTranslation
As a translation administrator, I every now and then want to take a look
at Special:Translations when deciding whether to fuzzy or not to fuzzy
(e.g. a typo that could have led to misunderstandings – if I see that
no translators misunderstood it, I don’t fuzzy, but if there are wrong
translations, I fuzzy). Currently this means going back to the page
itself, then going to Special:Translate, searching for the message in
question, and going to Special:Translations from there. With this patch,
it becomes a single click.

While I don’t have such a use case at hand, also add the same link to
deleted units for consistency.

Also for consistency, parse the new unit message as well.

Change-Id: Id6f042f34105db694a3ef2ab5f487e8abb6ec5d0