Only load VisualEditor integration when it is loaded
On CI, we would like to remove the recursive injection of extensions
dependencies. When Translate is a dependency, we are required to add
VisualEditor to fullfil the requirement that all dependencies are
defined (ResourcesTest::testValidDependencies).

Move the `ext.translate.ve` registration from `extension.json` to the
PHP `ResourceLoaderRegisterModules`. Only load it when VisualEditor is
loaded.

On CI, this saves us from having to add VisualEditor as a dependency of
extensions depending or having tests with Translate:

- AdvancedSearch
- CampaignEvents
- FundraisingTranslateWorkflow
- TranslationNotifications
- TwnMainPage

Bug: T406376
Bug: T412403
Change-Id: I0dbed826b7e486eae1143fed87737ecec9ec2123
(cherry picked from commit 62253e0ca5d01a0be678404eaf97b45c3192d19e)