tux: Don't display unusable shortcuts
When holding the Alt key, the translate interface shows shortcuts
for inserting various things. So when you continue to hold Alt,
then press the displayed number, it will insert that insertable
in the textarea.

However, you can't hold Alt and press two keys in succession, so
two-digit numbers should not be displayed. This brings the code
in ext.translate.editor.shortcuts.js (which *displays* the
shortcuts) more in line with the code in ext.translate.editor.js
(which adds the actual functionality when you use the shortcuts),
which already is limited to indexes between 0 and 9.

Bug: T406826
Change-Id: I6ecaa2f23b1099c1622f97ebd974997e9b300171