Redefine "isSourcePage" to include broken pages
To fix various bugs involving broken pages not being handled properly by
parts of the Translate UI once the source page cache expires.

As part of this, I reviewed all uses of the method to see whether they made sense for broken or unmarked pages.

It appears that the following outright bugs are fixed:

- Translation pages with a broken source can no longer be edited
  normally.
- Trying to move or delete a translatable page that is broken or a
  translation page with a broken source now goes through Translate's UI
  not core's.
- The "translate this page" link for broken pages no longer renders
  unstyled in an awkward manner.

The following debatable behavioral changes are made (I could be
convinced to go either way on some of these):
 - Transcluding a broken page now uses translation-aware transclusion if
   configured (and hence transcludes a variation of the last marked
   version)
 - Editing a broken page now shows the source page editing help message.
 - Broken pages do not show a "Mark for translation" link in the menu
   (this is explicitly filtered out in the code to show it)

Bug: T410746
Bug: T346587
Change-Id: Ibcdbf2a6f16b8c2fd3e2e1d9c62232139466f0c9